+++ ui-core/src/main/java/com/adyen/checkout/ui/core/internal/DefaultRedirectHandler.kt
@@ -14,13 +14,11 @@ import android.net.Uri
 import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.annotation.RestrictTo
-import androidx.browser.customtabs.CustomTabColorSchemeParams
-import androidx.browser.customtabs.CustomTabsIntent
 import com.adyen.checkout.core.AdyenLogLevel
 import com.adyen.checkout.core.exception.CheckoutException
 import com.adyen.checkout.core.exception.ComponentException
 import com.adyen.checkout.core.internal.util.adyenLog
-import com.adyen.checkout.ui.core.internal.util.ThemeUtil
+import com.adyen.checkout.ui.core.internal.util.CustomTabsLauncher
 import org.json.JSONException
 import org.json.JSONObject
 import java.lang.ref.WeakReference
@@ -153,25 +151,15 @@ class DefaultRedirectHandler : RedirectHandler {
 
     private fun launchWithCustomTabs(context: Context, uri: Uri): Boolean {
         // open in custom tabs if there's no native app for the target uri
-        val defaultColors = CustomTabColorSchemeParams.Builder()
-            .setToolbarColor(ThemeUtil.getPrimaryThemeColor(context))
-            .build()
-
-        @Suppress("SwallowedException")
-        return try {
-            CustomTabsIntent.Builder()
-                .setShowTitle(true)
-                .setDefaultColorSchemeParams(defaultColors)
-                .build()
-                .launchUrl(context, uri)
+        val isLaunched = CustomTabsLauncher.launchCustomTab(context, uri)
+        if (isLaunched) {
             adyenLog(AdyenLogLevel.DEBUG) { "launchWithCustomTabs - redirect successful with custom tabs" }
-            true
-        } catch (e: ActivityNotFoundException) {
+        } else {
             adyenLog(AdyenLogLevel.DEBUG) {
                 "launchWithCustomTabs - device doesn't support custom tabs or chrome is disabled"
             }
-            false
         }
+        return isLaunched
     }
 
     /**