+++ RELEASE_NOTES.md
@@ -1,12 +1,15 @@
 [//]: # (This file will be used for the release notes on GitHub when publishing.)
 [//]: # (Types of changes: `Breaking changes` `New` `Added` `Improved` `Changed` `Deprecated` `Removed` `Fixed`)
 [//]: # (Example:)
-[//]: # (## Added)
+[//]: # (## New)
 [//]: # ( - New payment method)
 [//]: # (## Changed)
 [//]: # ( - DropIn service's package changed from `com.adyen.dropin` to `com.adyen.dropin.services`)
 [//]: # (## Deprecated)
 [//]: # ( - Configurations public constructor are deprecated, please use each Configuration's builder to make a Configuration object)
+
+## New
+- For external redirects launched in a Custom Tab, you can now [customize the toolbar and navigation bar colors](docs/UI_CUSTOMIZATION.md#styling-custom-tabs) of the Custom Tab.
 
 ## Fixed
 - Fixed various memory leaks.